# docs/coach/coach-action.openapi.yaml
openapi: 3.1.0
info:
  title: Weight Control Coach Action
  description: Discovers and reads scoped Weight Control health data, manages confirmed nutrition records, health constraints, and coaching plans, and preserves the reflection workflow.
  version: 1.3.0
servers:
  - url: https://weightcontrol.devjllado.com/api/chatgpt-actions
paths:
...
  /coach/context:
    get:
      operationId: getHealthContext
      summary: Get a minimal domain-scoped health context
      description: Returns only the requested health domains for an inclusive range of no more than 90 days. Today is allowed even when the dashboard day is incomplete.
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive context start date in YYYY-MM-DD format.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive context end date in YYYY-MM-DD format. It cannot be in the future.
          schema:
            type: string
            format: date
        - name: domains
          in: query
          required: true
          description: One or more uppercase Coach domains as a comma-separated list.
          style: form
          explode: false
          schema:
            type: array
            minItems: 1
            uniqueItems: true
            items:
              $ref: "#/components/schemas/CoachDomain"
      responses:
        "200":
          description: Requested health context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachContext"
        "400":
          description: Missing domains, invalid dates, a future end date, or a range longer than 90 inclusive days
  /coach/meals:
    get:
      operationId: getMeals
      summary: List meals for review or editing
      description: Returns owned meals and their resource IDs for an inclusive range of no more than 90 days.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Stored meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
        "400":
          description: Invalid dates, a future end date, or a range longer than 90 inclusive days
    post:
      operationId: createMeal
      summary: Create a confirmed meal
      description: Creates the exact meal previously proposed to and explicitly confirmed by the user.
      x-openai-isConsequential: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealWrite"
      responses:
        "200":
          description: Created meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Invalid values or missing explicit confirmation
  /coach/meals/{id}:
    put:
      operationId: updateMeal
      summary: Update a confirmed meal
      description: Replaces one owned meal with the exact values explicitly confirmed by the user. Call getMeals first to obtain the resource ID.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealWrite"
      responses:
        "200":
          description: Updated meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Invalid values or missing explicit confirmation
        "404":
          description: Meal not found for the authenticated user
    delete:
      operationId: deleteMeal
      summary: Delete a confirmed meal
      description: Deletes one owned meal after the user explicitly confirms deleting the exact record. Call getMeals first to obtain the resource ID.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmedWrite"
      responses:
        "200":
          description: Meal deleted
        "400":
          description: Missing explicit confirmation
        "404":
          description: Meal not found for the authenticated user
  /coach/fasting-periods:
    get:
      operationId: getFastingPeriods
      summary: List fasting periods for review or editing
      description: Returns owned fasting periods and their resource IDs when they overlap an inclusive date range of no more than 90 days.
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Stored fasting periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FastingPeriod"
        "400":
          description: Invalid dates, a future end date, or a range longer than 90 inclusive days
    post:
      operationId: createFastingPeriod
      summary: Create a confirmed fasting period
      description: Creates the exact completed fasting period previously proposed to and explicitly confirmed by the user.
      x-openai-isConsequential: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FastingPeriodWrite"
      responses:
        "200":
          description: Created fasting period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastingPeriod"
        "400":
          description: Invalid values, an overlapping period, or missing explicit confirmation
  /coach/fasting-periods/{id}:
    put:
      operationId: updateFastingPeriod
      summary: Update a confirmed fasting period
      description: Replaces one owned fasting period with the exact values explicitly confirmed by the user. Call getFastingPeriods first to obtain the resource ID.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FastingPeriodWrite"
      responses:
        "200":
          description: Updated fasting period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastingPeriod"
        "400":
          description: Invalid values, an overlapping period, or missing explicit confirmation
        "404":
          description: Fasting period not found for the authenticated user
    delete:
      operationId: deleteFastingPeriod
      summary: Delete a confirmed fasting period
      description: Deletes one owned fasting period after the user explicitly confirms deleting the exact record. Call getFastingPeriods first to obtain the resource ID.
      x-openai-isConsequential: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmedWrite"
      responses:
        "200":
          description: Fasting period deleted
        "400":
          description: Missing explicit confirmation
        "404":
          description: Fasting period not found for the authenticated user
...
  /coach/active-plan:
    get:
      operationId: getActivePlan
      summary: Get the complete active coaching plan
      description: Returns the current goal, principles, ordered priorities, agreed actions, dates, notes, and update time.
      responses:
        "200":
          description: Complete active coaching plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingPlan"
        "204":
          description: The authenticated user has no active coaching plan
    put:
      operationId: updateActivePlan
      summary: Replace the complete active coaching plan after confirmation
      description: Creates or atomically replaces the complete plan with the exact version explicitly confirmed by the user.
      x-openai-isConsequential: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachingPlanWrite"
      responses:
        "200":
          description: Replaced active coaching plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingPlan"
        "400":
          description: Invalid values or missing explicit confirmation
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    CoachDomain:
      type: string
      enum:
        - PROFILE
        - BODY
        - VITALS
        - NUTRITION
        - TRAINING
        - RECOVERY
        - BEHAVIOR
        - HEALTH_EVENTS
        - HEALTH_CONSTRAINTS
        - ACTIVE_PLAN
        - DECISIONS
        - REFLECTIONS
...
    CoachContext:
      type: object
      additionalProperties: false
      required:
        - timezone
        - currentLocalDateTime
        - from
        - to
        - endDateComplete
        - dataSemantics
        - data
      properties:
        timezone:
          type: string
          example: Europe/Madrid
        currentLocalDateTime:
          type: string
          format: date-time
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        lastCompletedDate:
          type:
            - string
            - "null"
          format: date
        endDateComplete:
          type: boolean
          description: True when the requested end date is no later than the last completed dashboard date.
        dataSemantics:
          $ref: "#/components/schemas/CoachDataSemantics"
        data:
          $ref: "#/components/schemas/CoachContextData"
    CoachDataSemantics:
      type: object
      additionalProperties: false
      required:
        - absentRecordsAreUnknown
        - recordedZeroCaloriesAreValid
        - requestedRangeIsInclusive
      properties:
        absentRecordsAreUnknown:
          type: boolean
          const: true
        recordedZeroCaloriesAreValid:
          type: boolean
          const: true
        requestedRangeIsInclusive:
          type: boolean
          const: true
    CoachContextData:
      type: object
      additionalProperties: false
      description: Contains only the domains requested through the domains query parameter.
      properties:
        PROFILE:
          $ref: "#/components/schemas/Profile"
        BODY:
          $ref: "#/components/schemas/BodyContext"
        VITALS:
          $ref: "#/components/schemas/VitalsContext"
        NUTRITION:
          $ref: "#/components/schemas/NutritionContext"
        TRAINING:
          $ref: "#/components/schemas/WorkoutContext"
        RECOVERY:
          $ref: "#/components/schemas/RecoveryContext"
        BEHAVIOR:
          $ref: "#/components/schemas/BehaviorContext"
        HEALTH_EVENTS:
          $ref: "#/components/schemas/HealthEventsContext"
        HEALTH_CONSTRAINTS:
          $ref: "#/components/schemas/HealthConstraintsContext"
        ACTIVE_PLAN:
          $ref: "#/components/schemas/ActivePlanContext"
        DECISIONS:
          $ref: "#/components/schemas/DecisionsContext"
        REFLECTIONS:
          $ref: "#/components/schemas/ReflectionsContext"
    BodyContext:
      type: object
      additionalProperties: false
      required:
        - measurements
      properties:
        measurements:
          type: array
          items:
            $ref: "#/components/schemas/BodyMeasurement"
    BodyMeasurement:
      type: object
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
        weightKg:
          type: number
        fatPercentage:
          type: number
        fatMassKg:
          type: number
        muscleMassKg:
          type: number
        musclePercentage:
          type: number
        weightChangeKg:
          type: number
        fatChangeKg:
          type: number
        muscleChangeKg:
          type: number
    VitalsContext:
      type: object
      additionalProperties: false
      required:
        - bloodPressures
        - lipidPanels
      properties:
        bloodPressures:
          type: array
          items:
            $ref: "#/components/schemas/BloodPressure"
        lipidPanels:
          type: array
          items:
            $ref: "#/components/schemas/LipidPanel"
    NutritionContext:
      type: object
      additionalProperties: false
      required:
        - dailyTotals
        - meals
        - fastingPeriods
      properties:
        dailyTotals:
          type: array
          items:
            $ref: "#/components/schemas/NutritionDailyTotal"
        meals:
          type: array
          description: Meal context without resource IDs; call getMeals before editing.
          items:
            $ref: "#/components/schemas/NutritionMeal"
        fastingPeriods:
          type: array
          description: Fasting context without resource IDs; call getFastingPeriods before editing.
          items:
            $ref: "#/components/schemas/NutritionFastingPeriod"
...
    NutritionFastingPeriod:
      type: object
      additionalProperties: false
      required:
        - startTime
        - endTime
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type:
            - string
            - "null"
    Meal:
      type: object
      additionalProperties: false
      required:
        - id
        - dateFormat
        - date
        - mealType
        - mealSequence
        - calories
        - source
      properties:
        id:
          type: integer
          format: int64
        dateFormat:
          type: string
        date:
          type: string
          format: date
        mealType:
          $ref: "#/components/schemas/MealType"
        mealSequence:
          type: integer
          minimum: 1
        mealTime:
          type:
            - string
            - "null"
          format: time
        calories:
          type: integer
          minimum: 0
        proteinGrams:
          type:
            - number
            - "null"
        carbohydrateGrams:
          type:
            - number
            - "null"
        fatGrams:
          type:
            - number
            - "null"
        notes:
          type:
            - string
            - "null"
        source:
          $ref: "#/components/schemas/MealSource"
    MealWrite:
      type: object
      additionalProperties: false
      required:
        - date
        - mealType
        - calories
        - source
        - confirmed
      properties:
        date:
          type: string
          format: date
        mealType:
          $ref: "#/components/schemas/MealType"
        mealTime:
          type:
            - string
            - "null"
          format: time
        calories:
          type: integer
          minimum: 0
        proteinGrams:
          type:
            - number
            - "null"
          minimum: 0
        carbohydrateGrams:
          type:
            - number
            - "null"
          minimum: 0
        fatGrams:
          type:
            - number
            - "null"
          minimum: 0
        notes:
          type:
            - string
            - "null"
        source:
          $ref: "#/components/schemas/MealSource"
        confirmed:
          type: boolean
          const: true
          description: Must be true only after the user explicitly confirms these exact values.
    MealType:
      type: string
      enum:
        - BREAKFAST
        - LUNCH
        - DINNER
        - SNACK
    MealSource:
      type: string
      enum:
        - MANUAL
        - GPT_IMAGE_ESTIMATE
    FastingPeriod:
      type: object
      additionalProperties: false
      required:
        - id
        - startTimeFormat
        - endTimeFormat
        - startTime
        - endTime
      properties:
        id:
          type: integer
          format: int64
        startTimeFormat:
          type: string
        endTimeFormat:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type:
            - string
            - "null"
    FastingPeriodWrite:
      type: object
      additionalProperties: false
      required:
        - startTime
        - endTime
        - confirmed
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        notes:
          type:
            - string
            - "null"
        confirmed:
          type: boolean
          const: true
          description: Must be true only after the user explicitly confirms these exact values.
    ConfirmedWrite:
      type: object
      additionalProperties: false
      required:
        - confirmed
      properties:
        confirmed:
          type: boolean
          const: true
          description: Must be true only after the user explicitly confirms this exact deletion.
    RecoveryContext:
      type: object
      additionalProperties: false
      required:
        - moods
        - sleeps
      properties:
        moods:
          type: array
          items:
            $ref: "#/components/schemas/CoachMood"
        sleeps:
          type: array
          items:
            $ref: "#/components/schemas/Sleep"
...
    BehaviorContext:
      type: object
      additionalProperties: false
      required:
        - dailyStatuses
        - habits
        - routines
      properties:
        dailyStatuses:
          type: array
          items:
            $ref: "#/components/schemas/CoachDailyStatus"
        habits:
          type: array
          items:
            $ref: "#/components/schemas/Habit"
        routines:
          type: array
          items:
            $ref: "#/components/schemas/CoachRoutine"
...
    HealthEventsContext:
      type: object
      additionalProperties: false
      required:
        - sicknesses
        - backPainEpisodes
      properties:
        sicknesses:
          type: array
          items:
            $ref: "#/components/schemas/Sickness"
        backPainEpisodes:
          type: array
          description: Location-specific entries; multiple entries can share a date and period when their regions or sides differ.
          items:
            $ref: "#/components/schemas/BackPainEpisode"
    HealthConstraintsContext:
      type: object
      additionalProperties: false
      required:
        - constraints
      properties:
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/HealthConstraintContext"
...
    ActivePlanContext:
      type: object
      additionalProperties: false
      required:
        - plan
      properties:
        plan:
          oneOf:
            - $ref: "#/components/schemas/CoachingPlan"
            - type: "null"
    CoachingPlan:
      type: object
      additionalProperties: false
      required:
        - goal
        - principles
        - priorities
        - actions
        - startDate
        - reviewDate
        - notes
        - updatedAt
      properties:
        goal:
          type: string
        principles:
          $ref: "#/components/schemas/CoachingPlanItems"
        priorities:
          $ref: "#/components/schemas/CoachingPlanItems"
        actions:
          $ref: "#/components/schemas/CoachingPlanItems"
        startDate:
          type: string
          format: date
        reviewDate:
          type:
            - string
            - "null"
          format: date
          description: Review checkpoint; it does not expire the plan.
        notes:
          type:
            - string
            - "null"
        updatedAt:
          type: string
          format: date-time
    CoachingPlanWrite:
      type: object
      additionalProperties: false
      required:
        - goal
        - principles
        - priorities
        - actions
        - startDate
        - reviewDate
        - notes
        - confirmed
      properties:
        goal:
          type: string
          minLength: 1
          maxLength: 255
        principles:
          $ref: "#/components/schemas/CoachingPlanItems"
        priorities:
          $ref: "#/components/schemas/CoachingPlanItems"
        actions:
          $ref: "#/components/schemas/CoachingPlanItems"
        startDate:
          type: string
          format: date
        reviewDate:
          type:
            - string
            - "null"
          format: date
        notes:
          type:
            - string
            - "null"
        confirmed:
          type: boolean
          const: true
          description: Must be true only after the user explicitly confirms this exact complete plan.
    CoachingPlanItems:
      type: array
      items:
        type: string
        minLength: 1
...
    DecisionsContext:
      type: object
      additionalProperties: false
      required:
        - outcomes
        - summary
      properties:
        outcomes:
          type: array
          items:
            $ref: "#/components/schemas/CoachDecision"
        summary:
          $ref: "#/components/schemas/DecisionRangeSummary"
...
    ReflectionsContext:
      type: object
      additionalProperties: false
      required:
        - reflections
      properties:
        reflections:
          type: array
          items:
            $ref: "#/components/schemas/RecentReflection"
...
    Profile:
      type: object
      properties:
        age:
          type:
            - integer
            - "null"
        heightCm:
          type:
            - integer
            - "null"
        sex:
          type:
            - string
            - "null"
          enum:
            - MALE
            - FEMALE
            - null
        fitnessLevel:
          type:
            - string
            - "null"
          enum:
            - SEDENTARY
            - LOW_ACTIVE
            - ACTIVE
            - VERY_ACTIVE
            - null
        takesMedication:
          type: boolean
        weeklyAverageCalorieMaximum:
          type: integer
        typicalCaloriesByWeekday:
          type: object
          properties:
            MONDAY:
              type: integer
            TUESDAY:
              type: integer
            WEDNESDAY:
              type: integer
            THURSDAY:
              type: integer
            FRIDAY:
              type: integer
            SATURDAY:
              type: integer
            SUNDAY:
              type: integer
...
    BloodPressure:
      type: object
      properties:
        date:
          type: string
          format: date
        systolic:
          type:
            - integer
            - "null"
        diastolic:
          type:
            - integer
            - "null"
        systolicChange:
          type:
            - integer
            - "null"
        diastolicChange:
          type:
            - integer
            - "null"
    LipidPanel:
      type: object
      additionalProperties: false
      required:
        - date
        - totalCholesterol
        - hdlCholesterol
        - ldlCholesterol
        - triglycerides
      properties:
        date:
          type: string
          format: date
        totalCholesterol:
          type: integer
          minimum: 1
        hdlCholesterol:
          type: integer
          minimum: 1
        ldlCholesterol:
          type: integer
          minimum: 1
        triglycerides:
          type: integer
          minimum: 1
...
    WorkoutContext:
      type: object
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutDay"
        exerciseSummaries:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutExercise"
...
security:
  - BearerAuth: []